Changed asset management to be broken down to 4 steps (Asset management broken down to) Cache: AssetCache - Holds the loaders and AssetPools - Provides interface to assets Pool: AssetPool - TODO; Holds the actual assets Proxy: AssetFinder - Proxy to other classes to retrieve assets from the asset cache Loaders: AssetFactory - Creates / Loads the assets and spits them out to the relevant pool determined by the cache.
AssetFactory now take a AssetFactorLoader to init the factory, allows a cleaner way to dependency inject into the factories for the systems required for the assets to be loaded. (Still requires work to be 100% cleaner, but much better then before)

Working on the object pool next and create custom deleters for the shared_ptr for the Asset so that it can possibly unload it self
with the systems that are passed through with the AssetFactorLoaders.